Rework scope frame and node emission logic
Since node appending is more complicated for environment nodes than
for compound nodes, due to the fact that new child nodes may also be
appended to opening/closing subtrees, it seems reasonable to get rid of
children lists in scope frames and append child nodes to parents
immediately instead of when the scope is closed. This will enable
introduction of node emission handlers for various scenarios, including
handling of environment nodes or even processing a stream of nodes
instead of building a complete DOM tree.