Add handling of 'at' characters in command names
Depending on whether \makeatletter or \makeatother command is reached,
'at' characters can be treated either as a letter, or as a special
character. Although the added solution is not completely correct
in terms of LaTeX, but it lets us distinguish cases when \@example
should be parsed as two nodes (e.g. in email address), or as a single
command node.